refactor: split handler module into abi/adapter/run
Maintainer review noted that the monolithic handler.rs (~940 lines)
mixed ABI types, the FfiHandler adapter, helpers, and run entry
points, blurring trust boundaries. Split into:

  * handler/abi.rs    — public ABI types, setters, accessors
  * handler/adapter.rs — FfiHandler + translate_action + drain helpers
  * handler/run.rs    — sandlock_run_with_handlers entry points

handler/mod.rs re-exports the full public surface so external
consumers and integration tests continue to import from
`sandlock_ffi::handler::*` unchanged.

Resolves comment 3 from PR #44 review.
